// backend/src/presentation/dtos/ag_register/letter.dto.ts
export class LetterDto {
    registerId: number;
    reporterName: string;
    reporterLastname: string;
    reportedName: string;
    reportedLastname: string;
    date: Date;
    time: string;
    code: string;
    cite: number;
    codeReporter: string;
    perforation: string;
    drillHole: number;
    direction: string;
    reason: string;
    dagmes_name: string;
    carCoste: number;
    costerMaterials: number;
    workersCoste: number;
    utilsCoste: number;
    machineCoste: number;
    recurringCoste: number;
    waterCoste: number;
    totalDamageCost: number;
  }